Add files via upload
In this repository are 3 neural networks, the purpose of which is to demonstrate and teach the levels of complexity and operation of this technology. 
The first is a 2-layer neural network capable of solving linear-type problems. The second is a neural network with a more complex multilayered architecture, where the activation functions are present and able to solve non-linear problems. Finally, the third is a convolutional neural network that receives external video inputs and in real-time, specialized in facial detection and identification.